collegamento algoritmo A*, fix di alcuni bug per il movimento sulla mappa: spostato l'arrivo da casella a pixel preciso così da evitare che si fermasse al primo pixel della casella

+++ pg.h
@@ -1,4 +1,5 @@
 #include "Mappa.h"
+#include <math.h>
 #include <SFML/Graphics.hpp>
 #ifndef PG_H
 #define PG_H
@@ -9,7 +10,7 @@ class pg {
 public:
     sf::CircleShape character;
     explicit pg(float x, float y);
-    void MoveDirection(float x, float y);
+    void MoveDirection(int x, int y);
 
 };
